+++ src/eosio.wps.cpp
@@ -9,3 +9,4 @@
 #include "periods.cpp"
 #include "settings.cpp"
 #include "vote.cpp"
+#include "proposers.cpp"